添加订单操作和用户管理。

+++ src/main/java/com/agile/crowdfunding/service/impl/UserServiceImpl.java
@@ -163,8 +163,8 @@ public class UserServiceImpl implements UserService {
         return "success";
     }
 
-    public Long total() {
-        return userRepository.countByUserIdIsNot("-1");
+    public Integer total() {
+        return userRepository.countByUserIdIsNot("-1").intValue();
     }
